correct dynamic loading examples
The current example won't work, as it mistakenly shows the ROOKOUT_TARGET_PID env var being set on the line IN FRONT of the java command. So first I have proposed removing that.

But then I've added an example that instead passes the needed env vars as java -D system properties. And note that I've added the -DROOKOUT_LABELS because in my testing of things, this would not work without it. (The injection would happen, but the site would never show a successful "connected application".)